+++ src/frontend/src/features/rooms/livekit/prefabs/ControlBar/ControlBar.tsx
@@ -5,7 +5,6 @@ import { MobileControlBar } from './MobileControlBar'
 import { DesktopControlBar } from './DesktopControlBar'
 import { SettingsDialogProvider } from '../../components/controls/SettingsDialogContext'
 import { useIsMobile } from '@/utils/useIsMobile'
-import { usePersistentUserChoices } from '../../hooks/usePersistentUserChoices'
 
 /** @public */
 export type ControlBarControls = {
@@ -48,53 +47,16 @@ export interface ControlBarProps extends React.HTMLAttributes<HTMLDivElement> {
  * @public
  */
 export function ControlBar({ onDeviceError }: ControlBarProps) {
-  const {
-    saveAudioInputEnabled,
-    saveVideoInputEnabled,
-    saveAudioInputDeviceId,
-    saveVideoInputDeviceId,
-  } = usePersistentUserChoices()
-
-  const microphoneOnChange = React.useCallback(
-    (enabled: boolean, isUserInitiated: boolean) =>
-      isUserInitiated ? saveAudioInputEnabled(enabled) : null,
-    [saveAudioInputEnabled]
-  )
-
-  const cameraOnChange = React.useCallback(
-    (enabled: boolean, isUserInitiated: boolean) =>
-      isUserInitiated ? saveVideoInputEnabled(enabled) : null,
-    [saveVideoInputEnabled]
-  )
-
-  const barProps = {
-    onDeviceError,
-    microphoneOnChange,
-    cameraOnChange,
-    saveAudioInputDeviceId,
-    saveVideoInputDeviceId,
-  }
-
   const isMobile = useIsMobile()
-
   return (
     <SettingsDialogProvider>
       {isMobile ? (
-        <MobileControlBar {...barProps} />
+        <MobileControlBar onDeviceError={onDeviceError} />
       ) : (
-        <DesktopControlBar {...barProps} />
+        <DesktopControlBar onDeviceError={onDeviceError} />
       )}
     </SettingsDialogProvider>
   )
 }
 
-export interface ControlBarAuxProps {
-  onDeviceError: ControlBarProps['onDeviceError']
-  microphoneOnChange: (
-    enabled: boolean,
-    isUserInitiated: boolean
-  ) => void | null
-  cameraOnChange: (enabled: boolean, isUserInitiated: boolean) => void | null
-  saveAudioInputDeviceId: (deviceId: string) => void
-  saveVideoInputDeviceId: (deviceId: string) => void
-}
+export type ControlBarAuxProps = Pick<ControlBarProps, 'onDeviceError'>
